# src/routes/openapi/openapi.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"TrelloCard": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"desc": {
						"type": "string"
					},
					"due": {
						"type": "string",
						"nullable": true
					},
					"dueComplete": {
						"type": "boolean"
					},
					"idList": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"desc",
					"due",
					"dueComplete",
					"idList"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTaskRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"dueDate": {
						"type": "string"
					}
				},
				"required": [
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateMultipleTasksRequest": {
				"properties": {
					"tasks": {
						"items": {
							"$ref": "#/components/schemas/CreateTaskRequest"
						},
						"type": "array"
					}
				},
				"required": [
					"tasks"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "your-node-app-name",
		"version": "1.0.0",
		"description": "Describe your app here.",
		"contact": {
			"name": "author_name"
		}
	},
	"paths": {
...
		"/tasks": {
			"post": {
				"operationId": "CreateTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TrelloCard"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTaskRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListTodos",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"count": {},
												"tasks": {
													"items": {},
													"type": "array"
												},
												"success": {
													"type": "boolean"
												},
												"text": {
													"type": "string"
												}
											},
											"required": [
												"tasks",
												"success",
												"text"
											],
											"type": "object"
										},
										{
											"properties": {
												"text": {},
												"tasks": {
													"items": {
														"$ref": "#/components/schemas/TrelloCard"
													},
													"type": "array"
												},
												"count": {
													"type": "number",
													"format": "double"
												},
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"tasks",
												"count",
												"success"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": []
			}
		},
		"/tasks/bulk": {
			"post": {
				"operationId": "CreateMultipleTasks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"tasks": {
											"items": {},
											"type": "array"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"tasks",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMultipleTasksRequest"
							}
						}
					}
				}
			}
		},
		"/tasks/{taskId}/complete": {
			"put": {
				"operationId": "CompleteTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TrelloCard"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/tasks/{taskId}": {
			"delete": {
				"operationId": "DeleteTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}